eclipse-3.4M7 bump.
svn path=/java-overlay/; revision=6244

+++ dev-java/eclipse-ecj/eclipse-ecj-3.4_pre7.ebuild
@@ -6,7 +6,7 @@ inherit eutils java-pkg-2
 
 MY_PN="ecj"
 MY_PV="${PV/_pre/M}"
-DMF="S-${MY_PV}-200803301350"
+DMF="S-${MY_PV}-200805020100"
 S="${WORKDIR}"
 
 DESCRIPTION="Eclipse Compiler for Java"
@@ -16,10 +16,14 @@ SRC_URI="http://download.eclipse.org/eclipse/downloads/drops/${DMF}/${MY_PN}src-
 LICENSE="EPL-1.0"
 KEYWORDS="~amd64 ~ia64 ~ppc ~ppc64 ~x86 ~x86-fbsd"
 SLOT="3.4"
-IUSE=""
+IUSE="java6"
 
-RDEPEND=">=virtual/jre-1.4"
-DEPEND=">=virtual/jdk-1.4
+RDEPEND="
+	java6? ( >=virtual/jre-1.6 )
+	!java6? ( >=virtual/jre-1.4 )"
+DEPEND="
+	java6? ( >=virtual/jdk-1.6 )
+	!java6? ( >=virtual/jdk-1.4 )
 	sys-apps/findutils
 	app-arch/unzip"
 
@@ -31,36 +35,36 @@ src_unpack() {
 	rm -f org/eclipse/jdt/core/JDTCompilerAdapter.java
 	rm -fr org/eclipse/jdt/internal/antadapter
 
-	# what the heck...?! java6
-	rm -fr org/eclipse/jdt/internal/compiler/tool/ \
-		org/eclipse/jdt/internal/compiler/apt/
+	# java6
+	if ! use java6; then
+		rm -fr org/eclipse/jdt/internal/compiler/tool/ \
+			org/eclipse/jdt/internal/compiler/apt/
+	fi
 
 	# gcj feature
-	epatch "${FILESDIR}"/${P}-gcj.diff
+	epatch "${FILESDIR}"/${PN}-3.4_pre6-gcj.diff
 }
 
 src_compile() {
-	local javac="javac" java="java" jar="jar"
-
 	mkdir -p bootstrap
 	cp -a org bootstrap
 
 	einfo "bootstrapping ${MY_PN} with javac"
 
 	cd "${S}"/bootstrap
-	${javac} $(find org/ -name '*.java') || die "${MY_PN} bootstrap failed!"
+	javac $(find org/ -name '*.java') || die "${MY_PN} bootstrap failed!"
 
 	find org/ -name '*.class' -o -name '*.properties' -o -name '*.rsc' | \
-		xargs ${jar} cf ${MY_PN}.jar
+		xargs jar cf ${MY_PN}.jar
 
 	einfo "build ${MY_PN} with bootstrapped ${MY_PN}"
 
 	cd "${S}"
-	${java} -classpath bootstrap/${MY_PN}.jar \
-		org.eclipse.jdt.internal.compiler.batch.Main -encoding ISO-8859-1 org \
-		|| die "${MY_PN} build failed!"
+	java -classpath bootstrap/${MY_PN}.jar \
+		org.eclipse.jdt.internal.compiler.batch.Main $(java-pkg_javac-args) \
+		-nowarn -encoding ISO-8859-1 org || die "${MY_PN} build failed!"
 	find org/ -name '*.class' -o -name '*.properties' -o -name '*.rsc' | \
-		xargs ${jar} cf ${MY_PN}.jar
+		xargs jar cf ${MY_PN}.jar
 }
 
 src_install() {
